+++ src/Exception/BrowserConnectionFailed.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * @license see LICENSE
+ */
+
+namespace HeadlessChromium\Exception;
+
+class BrowserConnectionFailed extends \Exception
+{
+
+}